requirements.txt
python-telegram-bot==13.15
SQLAlchemy==2.0.23
python-dotenv==1.0.0
alembic==1.12.1